+++ src/mainwindow.cpp
@@ -1226,7 +1226,8 @@ void MainWindow::onRunFinished(int index, const QString &out, const QString &err
     if (!err.trimmed().isEmpty())
         log->error(head + tr("/stderr"), err);
     testcases->setOutput(index, out);
-    if (!out.isEmpty() && !testcases->expected(index).isEmpty())
+    if ((!out.isEmpty() && !testcases->expected(index).isEmpty()) ||
+        (SettingsHelper::isCheckOnTestcasesWithEmptyOutput() && exitCode == 0))
         checker->reqeustCheck(index, testcases->input(index), out, testcases->expected(index));
 }